// src/api/generated/model/platformLanguageResponse.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Amrita ICPC Coding Platform - Backend
 * A backend tool for Amrita ICPC Coding Platform
 * OpenAPI spec version: 1.0.0
 */

export interface PlatformLanguageResponse {
  id: number;
  name: string;
  slug: string;
  file_extension: string | null;
  monaco_language: string | null;
}